Add CI config using GitHub Actions (#36)
* Document running the tests

* Add CI config using GitHub Actions

* Extract Go version from go.mod

* Remove fallback Go version

This should avoid confusion since the default is likely to fall behind
and could result in unexpected bahavior.

* Log the found Go version

* Clean target directories before unpacking cache

* Update setup-go to v6

This allows us to simplify the workflow and use setup-go's built-in
version detection and Go-aware caching.

* Run CI only on push

Running on pull_request triggered a redundant build when I opened a PR:
one for the push and one for the PR.  It seems like running only on push
will work well for now.